prog

+++ src/constraint_writer.rs
@@ -137,7 +137,7 @@ impl ConstraintWriter {
 
     pub fn indent(&self) -> Result<(), WidthLimitExceededError> {
         self.with_buffer(|b| {
-            b.extend(std::iter::repeat_n(' ', self.constraints.indent.get()))
+            b.extend(std::iter::repeat_n(' ', self.constraints.indent.get() as usize))
         });
         self.check_width_constraints()
     }
